feat(core): finalize AuraChainAI integration and deployment readiness
Complete the pending AI governance/distribution modules, fix TypeScript and Hardhat execution blockers, and synchronize README/.env templates so the stack can be built and deployed consistently.

Made-with: Cursor

+++ aura-l3-contracts/contracts/KarmaGuardian.sol
@@ -16,7 +16,7 @@ contract KarmaGuardian {
 
     function updateKarma(address citizen, int256 delta) public onlyAIRegulator {
         int256 base = int256(karmaScore[citizen]);
-        karmaScore[citizen] = uint256(base + delta > 0 ? base + delta : 0);
+        karmaScore[citizen] = uint256(base + delta > 0 ? base + delta : int256(0));
     }
 
     function eligibleToVote(address citizen) public view returns(bool) {
